inventory-management-api/requirements.txt
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
django-filter==24.3
django-picklefield==3.2
django-q==1.3.9
django-taggit==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==3.5.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
types-python-dateutil==2.9.0.20240906
uritemplate==4.1.1
wcwidth==0.2.13
